Iterating
Moved another method from the facets class into the new scriblio class.

Moved the identification of taxonomies that can become facets from the actual registration of them

Simplified the options array with fewer keys, but made it more complex by filtering it twice. The second filter call is after the init hook when taxonomies are registered.

Added an option key to prevent registration of the default taxonomies.

Removed code that automatically registered the Facets class when the file was loaded. The class is now loaded from within the new Scriblio class and the $facets global is pretty much dead. There is a facets() global function, but that will trigger a deprecated function warning. There are changes throughout the code to eliminate the use of the $facets global.

The register_facet() method inside the facets class will now attempt to lazy load the class file from among class files inside the plugin. As before, custom facet classes should be loaded prior to calling the register_facet() method.